(fileQueue) Extract FileQueueItem subcomponent to its own module
